Add files via upload
Here we will write our own k-means model from scratch and use it to cluster handwritten numbers from the MNIST dataset

In the cell below, we create an assign_data function, which takes the data and the centers for each cluster and makes an assignment of each datapoint in data to the closest of the centers, centerids. We extract n, the number of datapoints, d, the dimensionality of the datapoints, and k the number of centers.

Next, we need to compute the squared distance between each center and each data point.

Reshaping the data to be 1 x n x d, and the centers to be k x 1 x d signals to numpy that when it subtracts these two arrays, it creates an array of shape k x n x d. That is, it computes all combinations of the k centers and the n datapoints for each of the d dimensions. We assign those differences to res.

Squaring each of the differences, then summing along dimension 2 --- that’s the components of the vectors --- produces the sum of squared distances, which is the squared distance between the centers and the datapoints. The resulting array is of shape k x n.

assign_data also computes the loss, which the sum of the squared differences. We want to know which center has the smallest squared distance for each data point. argmin produces the index of an array with the smallest value along the given dimension. Here, we’re using dimension 0, which varies over the k centers. centerids is now an array with one integer for each datapoint that indicates which of the centers is closest.